# Dedicated `ranges::minmax` vectorization that does not unnecessarily track element pointer