--- a/stl/inc/xutility
+++ b/stl/inc/xutility
@@ -108,6 +108,27 @@ const void* __stdcall __std_max_element_4(const void* _First, const void* _Last,
 const void* __stdcall __std_max_element_8(const void* _First, const void* _Last, bool _Signed) noexcept;
 const void* __stdcall __std_max_element_f(const void* _First, const void* _Last, bool _Unused) noexcept;
 const void* __stdcall __std_max_element_d(const void* _First, const void* _Last, bool _Unused) noexcept;
+
+__declspec(noalias) int8_t __stdcall __std_min_1i(const void* const _First, const void* const _Last) noexcept;
+__declspec(noalias) uint8_t __stdcall __std_min_1u(const void* const _First, const void* const _Last) noexcept;
+__declspec(noalias) int16_t __stdcall __std_min_2i(const void* const _First, const void* const _Last) noexcept;
+__declspec(noalias) uint16_t __stdcall __std_min_2u(const void* const _First, const void* const _Last) noexcept;
+__declspec(noalias) int32_t __stdcall __std_min_4i(const void* const _First, const void* const _Last) noexcept;
+__declspec(noalias) uint32_t __stdcall __std_min_4u(const void* const _First, const void* const _Last) noexcept;
+__declspec(noalias) int64_t __stdcall __std_min_8i(const void* const _First, const void* const _Last) noexcept;
+__declspec(noalias) uint64_t __stdcall __std_min_8u(const void* const _First, const void* const _Last) noexcept;
+__declspec(noalias) float __stdcall __std_min_f(const void* const _First, const void* const _Last) noexcept;
+__declspec(noalias) double __stdcall __std_min_d(const void* const _First, const void* const _Last) noexcept;
+__declspec(noalias) int8_t __stdcall __std_max_1i(const void* const _First, const void* const _Last) noexcept;
+__declspec(noalias) uint8_t __stdcall __std_max_1u(const void* const _First, const void* const _Last) noexcept;
+__declspec(noalias) int16_t __stdcall __std_max_2i(const void* const _First, const void* const _Last) noexcept;
+__declspec(noalias) uint16_t __stdcall __std_max_2u(const void* const _First, const void* const _Last) noexcept;
+__declspec(noalias) int32_t __stdcall __std_max_4i(const void* const _First, const void* const _Last) noexcept;
+__declspec(noalias) uint32_t __stdcall __std_max_4u(const void* const _First, const void* const _Last) noexcept;
+__declspec(noalias) int64_t __stdcall __std_max_8i(const void* const _First, const void* const _Last) noexcept;
+__declspec(noalias) uint64_t __stdcall __std_max_8u(const void* const _First, const void* const _Last) noexcept;
+__declspec(noalias) float __stdcall __std_max_f(const void* const _First, const void* const _Last) noexcept;
+__declspec(noalias) double __stdcall __std_max_d(const void* const _First, const void* const _Last) noexcept;
 } // extern "C"
 
 _STD_BEGIN
@@ -207,6 +228,96 @@ _Ty* __std_max_element(_Ty* _First, _Ty* _Last) noexcept {
         return const_cast<_Ty*>(static_cast<const _Ty*>(::__std_max_element_4(_First, _Last, _Signed)));
     } else if constexpr (sizeof(_Ty) == 8) {
         return const_cast<_Ty*>(static_cast<const _Ty*>(::__std_max_element_8(_First, _Last, _Signed)));
+    } else {
+        static_assert(_STD _Always_false<_Ty>, "Unexpected size");
+    }
+}
+
+template <class _Ty>
+auto __std_min(_Ty* _First, _Ty* _Last) noexcept {
+    constexpr bool _Signed = _STD is_signed_v<_Ty>;
+
+    if constexpr (_STD is_pointer_v<_Ty>) {
+        if constexpr (sizeof(_Ty) == 4) {
+            return reinterpret_cast<void*>(::__std_min_4u(_First, _Last));
+        } else if constexpr (sizeof(_Ty) == 8) {
+            return reinterpret_cast<void*>(::__std_min_8u(_First, _Last));
+        } else {
+            static_assert(_STD _Always_false<_Ty>, "Unexpected size");
+        }
+    } else if constexpr (_STD is_same_v<_STD remove_const_t<_Ty>, float>) {
+        return ::__std_min_f(_First, _Last);
+    } else if constexpr (_STD _Is_any_of_v<_STD remove_const_t<_Ty>, double, long double>) {
+        return ::__std_min_d(_First, _Last);
+    } else if constexpr (sizeof(_Ty) == 1) {
+        if constexpr (_Signed) {
+            return ::__std_min_1i(_First, _Last);
+        } else {
+            return ::__std_min_1u(_First, _Last);
+        }
+    } else if constexpr (sizeof(_Ty) == 2) {
+        if constexpr (_Signed) {
+            return ::__std_min_2i(_First, _Last);
+        } else {
+            return ::__std_min_2u(_First, _Last);
+        }
+    } else if constexpr (sizeof(_Ty) == 4) {
+        if constexpr (_Signed) {
+            return ::__std_min_4i(_First, _Last);
+        } else {
+            return ::__std_min_4u(_First, _Last);
+        }
+    } else if constexpr (sizeof(_Ty) == 8) {
+        if constexpr (_Signed) {
+            return ::__std_min_8i(_First, _Last);
+        } else {
+            return ::__std_min_8u(_First, _Last);
+        }
+    } else {
+        static_assert(_STD _Always_false<_Ty>, "Unexpected size");
+    }
+}
+
+template <class _Ty>
+auto __std_max(_Ty* _First, _Ty* _Last) noexcept {
+    constexpr bool _Signed = _STD is_signed_v<_Ty>;
+
+    if constexpr (_STD is_pointer_v<_Ty>) {
+        if constexpr (sizeof(_Ty) == 4) {
+            return reinterpret_cast<void*>(::__std_max_4u(_First, _Last));
+        } else if constexpr (sizeof(_Ty) == 8) {
+            return reinterpret_cast<void*>(::__std_max_8u(_First, _Last));
+        } else {
+            static_assert(_STD _Always_false<_Ty>, "Unexpected size");
+        }
+    } else if constexpr (_STD is_same_v<_STD remove_const_t<_Ty>, float>) {
+        return ::__std_max_f(_First, _Last);
+    } else if constexpr (_STD _Is_any_of_v<_STD remove_const_t<_Ty>, double, long double>) {
+        return ::__std_max_d(_First, _Last);
+    } else if constexpr (sizeof(_Ty) == 1) {
+        if constexpr (_Signed) {
+            return ::__std_max_1i(_First, _Last);
+        } else {
+            return ::__std_max_1u(_First, _Last);
+        }
+    } else if constexpr (sizeof(_Ty) == 2) {
+        if constexpr (_Signed) {
+            return ::__std_max_2i(_First, _Last);
+        } else {
+            return ::__std_max_2u(_First, _Last);
+        }
+    } else if constexpr (sizeof(_Ty) == 4) {
+        if constexpr (_Signed) {
+            return ::__std_max_4i(_First, _Last);
+        } else {
+            return ::__std_max_4u(_First, _Last);
+        }
+    } else if constexpr (sizeof(_Ty) == 8) {
+        if constexpr (_Signed) {
+            return ::__std_max_8i(_First, _Last);
+        } else {
+            return ::__std_max_8u(_First, _Last);
+        }
     } else {
         static_assert(_STD _Always_false<_Ty>, "Unexpected size");
     }
@@ -6971,6 +7082,13 @@ namespace ranges {
 _EXPORT_STD template <class _Ty, class _Pr>
 _NODISCARD constexpr _Ty(min)(initializer_list<_Ty> _Ilist, _Pr _Pred) {
     // return leftmost/smallest
+#if _USE_STD_VECTOR_ALGORITHMS
+    if constexpr (_Is_min_max_optimization_safe<const _Ty*, _Pr>) {
+        if (!_Is_constant_evaluated()) {
+            return static_cast<_Ty>(_STD __std_min(_To_address(_Ilist.begin()), _STD _To_address(_Ilist.end())));
+        }
+    }
+#endif // _USE_STD_VECTOR_ALGORITHMS
     const _Ty* _Res = _STD _Min_element_unchecked(_Ilist.begin(), _Ilist.end(), _STD _Pass_fn(_Pred));
     return *_Res;
 }
@@ -7004,6 +7122,13 @@ namespace ranges {
             const auto _Last  = _Range.end();
             _STL_ASSERT(_First != _Last,
                 "An initializer_list passed to std::ranges::min must not be empty. (N4950 [alg.min.max]/5)");
+#if _USE_STD_VECTOR_ALGORITHMS
+            if constexpr (_Is_min_max_optimization_safe<const _Ty*, _Pr>) {
+                if (!_Is_constant_evaluated()) {
+                    return static_cast<_Ty>(_STD __std_min(_STD _To_address(_First), _STD _To_address(_Last)));
+                }
+            }
+#endif // _USE_STD_VECTOR_ALGORITHMS
             return *_RANGES _Min_element_unchecked(_First, _Last, _STD _Pass_fn(_Pred), _STD _Pass_fn(_Proj));
         }
 
@@ -7017,6 +7142,16 @@ namespace ranges {
             _STL_ASSERT(
                 _UFirst != _ULast, "A range passed to std::ranges::min must not be empty. (N4950 [alg.min.max]/5)");
             if constexpr (forward_range<_Rng> && _Prefer_iterator_copies<iterator_t<_Rng>>) {
+#if _USE_STD_VECTOR_ALGORITHMS
+                if constexpr (is_same_v<_Pj, identity> && _Is_min_max_optimization_safe<decltype(_UFirst), _Pr>
+                              && sized_sentinel_for<decltype(_ULast), decltype(_UFirst)>) {
+                    if (!_STD is_constant_evaluated()) {
+                        const auto _First_ptr = _STD to_address(_UFirst);
+                        const auto _Last_ptr  = _First_ptr + (_ULast - _UFirst);
+                        return static_cast<range_value_t<_Rng>>(_STD __std_min(_First_ptr, _Last_ptr));
+                    }
+                }
+#endif // _USE_STD_VECTOR_ALGORITHMS
                 return static_cast<range_value_t<_Rng>>(*_RANGES _Min_element_unchecked(
                     _STD move(_UFirst), _STD move(_ULast), _STD _Pass_fn(_Pred), _STD _Pass_fn(_Proj)));
             } else {