--- a/stl/inc/xutility
+++ b/stl/inc/xutility
@@ -7141,17 +7141,17 @@ namespace ranges {
             auto _ULast  = _Uend(_Range);
             _STL_ASSERT(
                 _UFirst != _ULast, "A range passed to std::ranges::min must not be empty. (N4950 [alg.min.max]/5)");
-            if constexpr (forward_range<_Rng> && _Prefer_iterator_copies<iterator_t<_Rng>>) {
 #if _USE_STD_VECTOR_ALGORITHMS
-                if constexpr (is_same_v<_Pj, identity> && _Is_min_max_optimization_safe<decltype(_UFirst), _Pr>
-                              && sized_sentinel_for<decltype(_ULast), decltype(_UFirst)>) {
-                    if (!_STD is_constant_evaluated()) {
-                        const auto _First_ptr = _STD to_address(_UFirst);
-                        const auto _Last_ptr  = _First_ptr + (_ULast - _UFirst);
-                        return static_cast<range_value_t<_Rng>>(_STD __std_min(_First_ptr, _Last_ptr));
-                    }
+            if constexpr (is_same_v<_Pj, identity> && _Is_min_max_optimization_safe<decltype(_UFirst), _Pr>
+                          && sized_sentinel_for<decltype(_ULast), decltype(_UFirst)>) {
+                if (!_STD is_constant_evaluated()) {
+                    const auto _First_ptr = _STD to_address(_UFirst);
+                    const auto _Last_ptr  = _First_ptr + (_ULast - _UFirst);
+                    return static_cast<range_value_t<_Rng>>(_STD __std_min(_First_ptr, _Last_ptr));
                 }
+            }
 #endif // _USE_STD_VECTOR_ALGORITHMS
+            if constexpr (forward_range<_Rng> && _Prefer_iterator_copies<iterator_t<_Rng>>) {
                 return static_cast<range_value_t<_Rng>>(*_RANGES _Min_element_unchecked(
                     _STD move(_UFirst), _STD move(_ULast), _STD _Pass_fn(_Pred), _STD _Pass_fn(_Proj)));
             } else {