--- a/stl/inc/xutility
+++ b/stl/inc/xutility
@@ -6884,6 +6884,13 @@ namespace ranges {
 _EXPORT_STD template <class _Ty, class _Pr>
 _NODISCARD constexpr _Ty(max)(initializer_list<_Ty> _Ilist, _Pr _Pred) {
     // return leftmost/largest
+#if _USE_STD_VECTOR_ALGORITHMS
+    if constexpr (_Is_min_max_optimization_safe<const _Ty*, _Pr>) {
+        if (!_Is_constant_evaluated()) {
+            return static_cast<_Ty>(_STD __std_max(_To_address(_Ilist.begin()), _STD _To_address(_Ilist.end())));
+        }
+    }
+#endif // _USE_STD_VECTOR_ALGORITHMS
     const _Ty* _Res = _STD _Max_element_unchecked(_Ilist.begin(), _Ilist.end(), _STD _Pass_fn(_Pred));
     return *_Res;
 }
@@ -6923,6 +6930,13 @@ namespace ranges {
             const auto _Last  = _Range.end();
             _STL_ASSERT(_First != _Last,
                 "An initializer_list passed to std::ranges::max must not be empty. (N4950 [alg.min.max]/13)");
+#if _USE_STD_VECTOR_ALGORITHMS
+            if constexpr (_Is_min_max_optimization_safe<const _Ty*, _Pr>) {
+                if (!_Is_constant_evaluated()) {
+                    return static_cast<_Ty>(_STD __std_max(_STD _To_address(_First), _STD _To_address(_Last)));
+                }
+            }
+#endif // _USE_STD_VECTOR_ALGORITHMS
             return *_RANGES _Max_element_unchecked(_First, _Last, _STD _Pass_fn(_Pred), _STD _Pass_fn(_Proj));
         }
 
@@ -6935,6 +6949,16 @@ namespace ranges {
             auto _ULast  = _Uend(_Range);
             _STL_ASSERT(
                 _UFirst != _ULast, "A range passed to std::ranges::max must not be empty. (N4950 [alg.min.max]/13)");
+#if _USE_STD_VECTOR_ALGORITHMS
+            if constexpr (is_same_v<_Pj, identity> && _Is_min_max_optimization_safe<decltype(_UFirst), _Pr>
+                          && sized_sentinel_for<decltype(_ULast), decltype(_UFirst)>) {
+                if (!_STD is_constant_evaluated()) {
+                    const auto _First_ptr = _STD to_address(_UFirst);
+                    const auto _Last_ptr  = _First_ptr + (_ULast - _UFirst);
+                    return static_cast<range_value_t<_Rng>>(_STD __std_max(_First_ptr, _Last_ptr));
+                }
+            }
+#endif // _USE_STD_VECTOR_ALGORITHMS
             if constexpr (forward_range<_Rng> && _Prefer_iterator_copies<iterator_t<_Rng>>) {
                 return static_cast<range_value_t<_Rng>>(*_RANGES _Max_element_unchecked(
                     _STD move(_UFirst), _STD move(_ULast), _STD _Pass_fn(_Pred), _STD _Pass_fn(_Proj)));