--- a/tests/std/include/test_min_max_element_support.hpp
+++ b/tests/std/include/test_min_max_element_support.hpp
@@ -107,5 +107,15 @@ void test_case_min_max_element(const std::vector<T>& input) {
     assert(expected_max == actual_max_sized_range);
     assert(expected_minmax.first == actual_minmax_sized_range.min);
     assert(expected_minmax.second == actual_minmax_sized_range.max);
+
+    if (input.begin() != input.end()) {
+        auto actual_min_value    = std::ranges::min(input);
+        auto actual_max_value    = std::ranges::max(input);
+        auto actual_minmax_value = std::ranges::minmax(input);     
+        assert(*expected_min == actual_min_value);
+        assert(*expected_max == actual_max_value);
+        assert(*expected_minmax.first == actual_minmax_value.min);
+        assert(*expected_minmax.second == actual_minmax_value.max);
+    }
 #endif // _HAS_CXX20
 }