… stl/src/vector_algorithms.cpp