@StephanTLavavej 2024-02-12
Thanks! I resolved the remaining feedback comments, and decided to simply require SSE 4.2 for the new optimization, consistent with the previous optimization. This is a *lot* easier to reason about.

* Don't need `_To_address` for `initializer_list`.
* `_STL_ASSERT` non-empty inputs, update citations.
* Cleanup comment punctuation.
* Avoid risk by always guarding with `_Use_sse42()`.
* Make room for comments by renaming lambda `_First` and `_Second` to `_Val1` and `_Val2`.
  + Regex renamed: `\[\]\((__m128[id]?) _First, \1 _Second\) \{ return (\w+)\(_First, _Second\); \}`
  + To: `[]($1 _Val1, $1 _Val2) { return $2(_Val1, _Val2); }`
* Comment all `_Minmax_traits_MEOW` intrinsics above SSE2.
  + Remove comments from `_mm_min_epi16` and `_mm_max_epi16` - they're SSE2, not SSE4.1.