@StephanTLavavej 2024-02-15
I'm *speculatively* mirroring this to the MSVC-internal repo - please notify me if any further changes are pushed.